refactor : 함수 폴더 위치 변경 및 변수명 수정
